Use CGFloats

+++ Canvas/DrawingStrategies/CVSBoxDrawingStrategy.m
@@ -23,7 +23,7 @@
     
     self.lastUpdatedArea = CGRectZero;
     
-    float const lineWidth = 2.0f * [UIScreen mainScreen].scale;
+    CGFloat const lineWidth = 2.0f * [UIScreen mainScreen].scale;
     
     CGContextSetLineWidth(context, lineWidth);
