+++ Canvas/UIColor+CVSExtensions.m
@@ -12,9 +12,9 @@
 
 + (UIColor*) cvs_randomColor
 {
-    float randomRed     = (arc4random() % 101)/100.0f;
-    float randomGreen   = (arc4random() % 101)/100.0f;
-    float randomBlue    = (arc4random() % 101)/100.0f;
+    CGFloat randomRed     = (arc4random() % 101)/100.0f;
+    CGFloat randomGreen   = (arc4random() % 101)/100.0f;
+    CGFloat randomBlue    = (arc4random() % 101)/100.0f;
     
     UIColor* color = [UIColor colorWithRed: randomRed green: randomGreen blue: randomBlue alpha:1.0f];
